Add ConfD

+++ validators/regexp/test.sh
@@ -5,7 +5,10 @@ RESULTSDIR=$ROOT_DIR/results/regexp
 OUTFILE=$RESULTSDIR/out
 FAILFILE=$RESULTSDIR/fail
 
-mkdir -p $RESULTSDIR
+if ! stat $RESULTSDIR; then
+  exit 0
+fi
+
 if go test -v gotests/regexp > $OUTFILE 2> $FAILFILE; then
   # Delete fail file if it's empty and the script passed.
   find $FAILFILE -size 0 -delete
